requirements/base.txt
celery
django<=2.0.7
django-compressor
django-hamlpy
django-redis
django-timezone-field
enum34
Pillow
phonenumbers
psycopg2==2.7.1
python-dateutil
rapidpro-python
redis
requests
smartmin
sorl-thumbnail
xlrd
xlwt
